cport: Teensy RAM1 fit + printf portability + gen_sd_pack self-sufficiency
First real Arduino IDE build attempt (user hardware) failed at the
teensy_size step: on the Teensy 4.x ALL code is copied into RAM1/ITCM
by default, and the engine (~200 KB of code) plus its ~170 KB of .bss —
dominated by the 78 KB framebuffer struct — plus the 80 KB save buffer
overflowed the 512 KB bank.  rd_state RD moves to RAM2/OCRAM
(.dmabuffers, guarded by __IMXRT1062__; safe because rd_init memsets
the whole struct — the section is not zeroed at boot) and the sketch's
savbuf gets DMAMEM, freeing ~160 KB of RAM1.

The -Wformat warnings the same build surfaced are fixed at their six
sites: int32_t is long on ARM, so the gold/substitution/rng prints use
%ld/%lu with explicit casts (host output unchanged).

gen_sd_pack.py now auto-runs port/tools/build_assets.py when
port/assets/manifest.json is missing (both are git-ignored regenerable
output, so a fresh checkout crashed on json.load) with a clear Pillow
hint on failure; the sketch README carries the full SD-prep chain.
Verified: the auto-built manifest reproduces the byte-identical PAK.

Host regression: make test, map/colony/europe render compares, input
savraleigh all green.

+++ cport/arduino/colopy/colopy_dialog.c
@@ -69,7 +69,7 @@ static void fill_template(const char *line, const rm_subs *subs,
                     while (*c && o + 1 < cap) out[o++] = *c++;
                 } else if (is_n && subs->num_set[k]) {
                     char nb[16];
-                    snprintf(nb, sizeof(nb), "%d", subs->num[k]);
+                    snprintf(nb, sizeof(nb), "%ld", (long)subs->num[k]);
                     for (const char *c = nb; *c && o + 1 < cap;)
                         out[o++] = *c++;
                 }
